stock_analysis/requirements.txt
lxml>=4.9.1
multitasking>=0.0.7
numpy>=1.16.5
pandas>=1.3.0
python-dotenv==1.0.0
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai>=0.0.5,<0.2.0
openai==1.23.6
langchain==0.1.16
unstructured~=0.10.25
crewai~=0.30.11
crewai_tools
tools~=0.1.9
agentops==0.2.0
reportlab
sec-api==1.0.18
requests==2.31.0
